Change Labeler to support more complex taxonomy
We now have multiple non-orthogonal label groups, as well as Severity
measures for some of the classes.

+++ lib/Video/Decode.cpp
@@ -136,7 +136,7 @@ Error VideoFile::SeekToFraction(double fraction_0_to_1) {
 }
 
 Error VideoFile::SeekToSecond(double second) {
-	return SeekToMicrosecond(second * 1000000.0);
+	return SeekToMicrosecond((int64_t)(second * 1000000.0));
 }
 
 Error VideoFile::SeekToMicrosecond(int64_t microsecond) {
